% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qHTSWaterfall.R
\name{plotWaterfall}
\alias{plotWaterfall}
\title{Plots 3D qHTS waterfall plot, given Pubchem activity file or NCATS qHTS format file.}
\usage{
plotWaterfall(
  inputFile,
  fileFormat = "generic_qhts",
  activityReadouts = c("Activity"),
  logMolarConcVector = NULL,
  pointColors = c("darkgreen", "blue4"),
  curveColors = c("darkgreen", "blue4"),
  inactiveColor = "gray",
  alpha = 1,
  pointSize = 1,
  lineWeight = 1.5,
  plotInactivePoints = T,
  curveResolution = 25,
  plotAspectRatio = c(1, 1, 3),
  antialiasSmoothing = F,
  axisTitles = list(concTitle = "log2[conc], M", respTitle = "Response", curveTitle =
    "Compound"),
  planeColors = list(basePlaneColor = "#b8b6b6", rightPlaneColor = "#999494",
    leftPlaneColor = "#6e6868"),
  gridColor = "#ffffff",
  showCurveNumberLabels = TRUE,
  concAxisConfig = NULL,
  responseAxisConfig = NULL,
  axisFontSize = 13,
  plotExportScale = 4,
  plotExportFileType = "png"
)
}
\arguments{
\item{inputFile}{The input file path.}

\item{fileFormat}{a required value indicdating the file format. Valid values are 'Generic_qhts' or 'NCATS_qhts'. The value is case-insensitive.}

\item{activityReadouts}{Activity data readouts to include in plot}

\item{logMolarConcVector}{An optional input list of log molar concentrations, if input is not from PubChem.}

\item{pointColors}{color list for activity readout \emph{\bold{data points}}.
The order should match the activityReadoutsList order.}

\item{curveColors}{color list for activity readout \emph{\bold{fitted dose-response curves}}.
The order should match the activityReadoutsList order.}

\item{inactiveColor}{color to display inactive data, default is 'gray'.}

\item{alpha}{alpha transparency of the the plot lines, default is 1.0}

\item{pointSize}{relative size of plotted points, default = 2}

\item{lineWeight}{thickness of fitted curves. Default thickness is 1.0. Decimal numbers are permitted.}

\item{plotInactivePoints}{TRUE will plot inactive data as datapoints, FALSE Will hide inactive data.}

\item{curveResolution}{value between 25 and 250, number of points to define dose-response curves.
Fewer points renders as connected straight lines.}

\item{plotAspectRatio}{relative sizes of concentration axis (x), response axis (y), and waterfall width (z).
Input as list, derault: c(1, 1, 3)}

\item{antialiasSmoothing}{smooths plot line rendering. Default is FALSE. Setting as TRUE will smooth lines, but may slow response during re-drawing plot.}

\item{axisTitles}{a list holding response, concentration and curve number axis titles (default, list(concTitle="log2Conc, M", respTitle="Response", curveTitle=""))}

\item{planeColors}{a list holding base, left and right plane colors, value names/keys are 'basePlaneColor, rightPlaneColor, leftPlaneColor.}

\item{gridColor}{a specified color for grids, color name or hex color value (default #ffffff)}

\item{showCurveNumberLabels}{boolean, indicates if the curve index numbers should be shown on the axis, or hidden.}

\item{concAxisConfig}{a list with 4 named values: min, max, tickWidth, firstTick.}

\item{responseAxisConfig}{a list with 4 named values: min, max, tickWidth, firstTick.}

\item{axisFontSize}{an integer specifying the font size for axis titles and tick labels. Default = 13, best size range tends to be from 11 to 15.}

\item{plotExportScale}{determines the export file size, higher results in increased relative resolution (range: 1-8), default is 4, 7 will give a result that is roughly 10MB}

\item{plotExportFileType}{default is png, other options include svg, jpeg or webp.}
}
\description{
Plots 3D qHTS waterfall plot, given Pubchem activity file or NCATS qHTS format file.
}
\examples{
\dontrun{

# specify sample data file
filePath <- system.file("extdata", "NCATS_CMT1A_PMP22_Follow_Up.csv", package="qHTSWaterfall")

# supply log molar concentrations (for non-PubChem data only)
logConc <- c(
 -9.011761791,
 -8.534640544,
 -8.057519287,
 -7.580398031,
 -7.103276777,
 -6.626155522,
 -6.149034267,
 -5.671913012,
 -5.194791758,
 -4.717670503,
 -4.240549248
)

# make plotting call with desired parameter settings
# 3D Plot will be presented in a separate window.
plotWaterfall(
  inputFile=filePath,
  fileFormat = 'ncats_qhts',
  activityReadouts = c("fluc", "nluc"),
  logMolarConcVector = logConc,
  pointColors = c("darkgreen", "royalblue3"),
  curveColors = c("darkgreen", "royalblue3"),
  inactiveColor = "gray",
  pointSize = 3,
  alpha = 1,
  plotInactivePoints = F,
  curveResolution = 100,
  plotAspectRatio = c(2,2,5),
  returnPlotObject = F
  )
}
}
