% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/3d_qHTS_pubchem_package.R
\name{plotWaterfall2}
\alias{plotWaterfall2}
\title{Plots 3D qHTS waterfall plot, given Pubchem activity file or NCATS qHTS format file.}
\usage{
plotWaterfall2(
  inputFile,
  activityReadouts = c("Activity"),
  logMolarConcVector,
  pointColors = c("darkgreen", "royalblue3"),
  curveColors = c("darkgreen", "royalblue3"),
  inactiveColor = "gray",
  alpha = 1
)
}
\arguments{
\item{inputFile}{An optional input file path. If none is provided, a file chooser will prompt.}

\item{activityReadouts}{Activity data readouts to include in plot}

\item{logMolarConcVector}{An optional input list of log molar concentrations, if input is not from PubChem.}

\item{pointColors}{color list for activity readout \emph{\bold{data points}}. The order should match the activityReadoutsList order.}

\item{curveColors}{color list for activity readout \emph{\bold{fitted dose-response curves}}. The order should match the activityReadoutsList order.}

\item{inactiveColor}{color to display inactive data, default is 'gray'.}

\item{alpha}{alpha transparency of the the plot lines, default is 1.0}
}
\description{
Plots 3D qHTS waterfall plot, given Pubchem activity file or NCATS qHTS format file.
}
\examples{
\dontrun{
plotWaterfall2(inputFile="./data/NCATS_qHTS_NPAC_Genesis_F1.csv", activityReadouts=c("Active", "Nluc100"))
}
}
