% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/runWaterfallApp.R
\name{runQHTSWaterfallApp}
\alias{runQHTSWaterfallApp}
\title{Runs the qHTSWaterfall Shiny Application User Interface locally}
\usage{
runQHTSWaterfallApp()
}
\description{
Runs the qHTSWaterfall Shiny Application User Interface locally
}
\examples{
\dontrun{
library(qHTSWaterfall)
runQHTSWaterfallApp()

# use 'Esc' button to exit application.

}
}
